+++ packages/core-data/src/hooks/index.ts
@@ -16,6 +16,7 @@ export {
 	default as useResourcePermissions,
 	useDeprecatedResourcePermissions as __experimentalUseResourcePermissions,
 } from './use-resource-permissions';
+export { default as useCoalescedEntityEdit } from './use-coalesced-entity-edit';
 export { default as useEntityBlockEditor } from './use-entity-block-editor';
 export { default as useEntityId } from './use-entity-id';
 export { default as useEntityProp } from './use-entity-prop';